+++ include/nanobind/stl/unique_ptr.h
@@ -96,15 +96,23 @@ struct type_caster<std::unique_ptr<T, Deleter>> {
         if constexpr (IsNanobindDeleter)
             cpp_delete = value.get_deleter().owned_by_cpp();
 
-        T *value_p = value.get();
+        T *ptr = value.get();
+        const std::type_info *type = &typeid(T);
+
+        constexpr bool has_type_hook =
+            !std::is_base_of_v<std::false_type, type_hook<T>>;
+        if constexpr (has_type_hook)
+            type = type_hook<T>::get(ptr);
 
         handle result;
-        if constexpr (std::is_polymorphic_v<T>)
-            result = nb_type_put_unique_p(&typeid(T),
-                                          value_p ? &typeid(*value_p) : nullptr,
-                                          value_p, cleanup, cpp_delete);
-        else
-            result = nb_type_put_unique(&typeid(T), value_p, cleanup, cpp_delete);
+        if constexpr (!std::is_polymorphic_v<T>) {
+            result = nb_type_put_unique(type, ptr, cleanup, cpp_delete);
+        } else {
+            const std::type_info *type_p =
+                (!has_type_hook && ptr) ? &typeid(*ptr) : nullptr;
+
+            result = nb_type_put_unique_p(type, type_p, ptr, cleanup, cpp_delete);
+        }
 
         if (result.is_valid()) {
             if (cpp_delete)