+++ tests/test_holders.py
@@ -158,8 +158,21 @@ def test07_polymorphic_downcast_unique():
     assert isinstance(t.u_polymorphic_factory_2(), t.PolymorphicBase)
 
 
-def test08_polymorphic_downcast_unique():
+def test08_polymorphic_downcast_shared():
     assert isinstance(t.s_factory(), t.Base)
     assert isinstance(t.s_factory_2(), t.Base)
     assert isinstance(t.s_polymorphic_factory(), t.PolymorphicSubclass)
     assert isinstance(t.s_polymorphic_factory_2(), t.PolymorphicBase)
+
+def test09_tag_based():
+    assert isinstance(t.make_pet(t.PetKind.Dog), t.Dog)
+    assert isinstance(t.make_pet(t.PetKind.Cat), t.Cat)
+
+
+def test09_tag_based_unique():
+    assert isinstance(t.make_pet_u(t.PetKind.Dog), t.Dog)
+    assert isinstance(t.make_pet_u(t.PetKind.Cat), t.Cat)
+
+def test09_tag_based_shared():
+    assert isinstance(t.make_pet_s(t.PetKind.Dog), t.Dog)
+    assert isinstance(t.make_pet_s(t.PetKind.Cat), t.Cat)